--- retrieval/query_expansion_prompt.py
@@ -1,222 +1,0 @@
-from textwrap import dedent
-
-
-ALLOWED_VALUES_GUIDE = dedent(
-    """
-    [허용 enum 값]
-    domain_scope: residential_lease | commercial_lease | mixed | unknown
-    source_type: law | case | counsel
-    priority: high | medium | low
-    confidence: high | medium | low
-
-    issue_types_normalized(한글 문자열만 사용):
-    - 보증금 반환·보전
-    - 갱신·해지·계약종료
-    - 차임·월세·보증금 증액
-    - 대출·계약금·조건부 해제
-    - 중도퇴거·중개보수·관리비
-    - 매매·신탁·등기·임대인지위승계
-    - 전입신고·확정일자·대항력
-    - 하자·수선·원상회복
-    - 상가·권리금·시설비
-    - 사용제한·반려동물·용도제한
-    - 약정금·지급명령·금전채권
-    - 기타
-    """
-).strip()
-
-
-TARGET_FIELDS_GUIDE = dedent(
-    """
-    [target_fields 허용값]
-    - law_query.target_fields: law_child_text, law_parent_text
-    - case_query.target_fields: case_issue_summary, case_holding_summary, case_referenced_law, case_full_text
-    - counsel_query.target_fields: counsel_question, counsel_tags, counsel_answer
-    """
-).strip()
-
-
-DOMAIN_KEYWORD_GUIDE = dedent(
-    """
-    [임대차 상담·특약 데이터 기반 검색 seed 표현]
-    - 보증금 반환, 임대차 계약, 계약 종료, 해지 통보, 묵시적 갱신
-    - 계약갱신요구권, 갱신거절, 실거주, 차임 증액, 월세 인상
-    - 전입신고, 확정일자, 대항력, 우선변제권, 최우선변제
-    - 임차권등기명령, 지급명령, 보증금반환청구
-    - 중도퇴거, 중개보수, 관리비, 원상회복, 수선의무, 하자
-    - 대출 불승인, 전세자금대출, 계약금 반환, 조건부 해제
-    - 매매계약, 신탁등기, 임대인 지위승계, 권리금, 시설비
-    """
-).strip()
-
-
-SYNONYM_GUIDE = dedent(
-    """
-    [생활어-법률어 확장 예시]
-    - 집주인 -> 임대인
-    - 세입자 -> 임차인
-    - 나간다, 퇴실 -> 퇴거, 명도, 인도
-    - 자동연장 -> 묵시적 갱신
-    - 보증금을 안 줌 -> 보증금 반환 지체, 임대차보증금 반환청구
-    - 전세대출 안 나옴 -> 대출 불승인, 조건 불성취, 계약금 반환
-    - 집 하자 -> 수선의무, 사용·수익 방해, 하자 보수
-    """
-).strip()
-
-
-OUTPUT_JSON_SKELETON_GUIDE = dedent(
-    """
-    [출력 JSON 구조 예시 - 필드 이름/타입을 반드시 따른다]
-    {
-      "schema_version": "qe_v1",
-      "clause_text": "입력 특약 원문",
-      "normalized_clause": "검색 친화적으로 정리한 문장",
-      "domain_scope": "residential_lease",
-      "issue_type_candidates_freeform": ["자유형 쟁점"],
-      "issue_types_normalized": ["전입신고·확정일자·대항력"],
-      "risk_hypotheses": ["검색할 위험 가설"],
-      "compound_clause_detected": false,
-      "sub_issue_summaries": [],
-      "law_query": {
-        "legal_issue": "문자열",
-        "applicable_rules": ["문자열 리스트"],
-        "law_article_candidates": [],
-        "law_keywords": ["전입신고", "확정일자", "대항력"],
-        "law_dense_query": "문장형 질의",
-        "retrieval_terms": {
-          "must_terms": ["전입신고"],
-          "should_terms": ["확정일자", "대항력"],
-          "exclude_terms": [],
-          "synonyms": ["주민등록"],
-          "query_variants": ["전입신고 제한 특약 관련 검색 질의"]
-        },
-        "target_fields": ["law_child_text"]
-      },
-      "case_query": {
-        "case_issue_query": "문자열",
-        "case_fact_pattern_query": "문자열",
-        "case_keywords": ["전입신고", "확정일자", "특약"],
-        "referenced_law_candidates": [],
-        "retrieval_terms": {
-          "must_terms": ["전입신고"],
-          "should_terms": ["확정일자", "대항력"],
-          "exclude_terms": [],
-          "synonyms": ["주민등록"],
-          "query_variants": ["전입신고 제한 특약 관련 판례 검색"]
-        },
-        "target_fields": ["case_issue_summary", "case_holding_summary"]
-      },
-      "counsel_query": {
-        "counsel_question_query": "문자열",
-        "counsel_answer_query": "문자열",
-        "user_question_intent": "문자열",
-        "counsel_keywords": ["전입신고", "확정일자", "임대차"],
-        "expected_tags": ["#임대차"],
-        "expected_answer_points": ["대항력", "우선변제권"],
-        "retrieval_terms": {
-          "must_terms": ["전입신고"],
-          "should_terms": ["확정일자", "보증금"],
-          "exclude_terms": [],
-          "synonyms": ["주민등록"],
-          "query_variants": ["전입신고 제한 특약 상담사례 검색"]
-        },
-        "target_fields": ["counsel_question", "counsel_answer"]
-      },
-      "source_routing": [],
-      "expansion_notes": null
-    }
-
-    금지:
-    - legal_issue, law_dense_query, case_issue_query, counsel_question_query 같은 문자열 필드에 객체를 넣지 않는다.
-    - query_text, article_no_candidates 같은 임의 필드를 만들지 않는다.
-    - source_routing에 confidence 필드를 넣지 않는다.
-    """
-).strip()
-
-
-SYSTEM_PROMPT = dedent(
-    f"""
-    너는 대한민국 임대차 계약서 특약 조항을 법령·판례·상담사례 검색용 질의로 변환하는
-    legal query expansion model이다.
-
-    너의 출력은 반드시 제공된 ClauseQueryExpansion Pydantic schema를 따른다.
-    너는 최종 법률 판단을 내리지 않는다.
-    너의 역할은 검색 성능을 높이기 위한 중간 산출물을 생성하는 것이다.
-
-    [출력 형식 절대 규칙]
-    - 응답은 순수 JSON 객체 하나만 출력한다.
-    - 코드블록, 설명문, 주석을 출력하지 않는다.
-    - 스키마에 없는 임의 필드를 만들지 않는다.
-
-    [핵심 원칙]
-    1) 최종 판단 금지
-    - "무효다", "위법이다", "반드시 위험하다"처럼 단정하지 않는다.
-    - risk_hypotheses는 검색할 위험 가설만 작성한다.
-
-    2) 원문 보존
-    - clause_text에는 입력 특약을 그대로 보존한다.
-    - normalized_clause에는 검색 친화적으로 정리한 문장을 쓴다.
-
-    3) source 분리
-    - law_query, case_query, counsel_query를 각각 작성한다.
-    - 불확실하면 후보 리스트는 빈 배열로 둬도 된다.
-
-    4) 조문번호 규칙
-    - article_no는 문자열 또는 null로 작성한다.
-    - 예: "3", "3의2", null
-
-    5) 쟁점 유형
-    - issue_type_candidates_freeform에는 자유형 쟁점을 쓴다.
-    - issue_types_normalized에는 허용된 한글 enum 문자열만 쓴다.
-
-    6) 복합 특약 처리
-    - 하나의 특약에 여러 쟁점이 있으면 compound_clause_detected를 true로 둔다.
-    - sub_issue_summaries에는 분리 가능한 하위 쟁점을 짧게 작성한다.
-
-    7) domain_scope 판단
-    - 주택임대차 중심이면 residential_lease
-    - 상가임대차 중심이면 commercial_lease
-    - 주택과 상가가 함께 있으면 mixed
-    - 판단이 어려우면 unknown
-
-    8) retrieval_terms 작성
-    - retrieval_terms는 핵심 표현 위주로 짧게 작성한다.
-    - 제외어가 없으면 exclude_terms는 빈 배열 []로 둔다.
-    - source_routing은 특별한 사유가 없으면 빈 배열 []로 둔다.
-    - source_routing을 작성할 때는 source_type, priority, reason만 사용한다.
-
-    {ALLOWED_VALUES_GUIDE}
-
-    {TARGET_FIELDS_GUIDE}
-
-    {OUTPUT_JSON_SKELETON_GUIDE}
-
-    {DOMAIN_KEYWORD_GUIDE}
-
-    {SYNONYM_GUIDE}
-    """
-).strip()
-
-
-def build_user_prompt(clause_text: str) -> str:
-    return dedent(
-        f"""
-        다음 임대차 계약서 특약 조항을 ClauseQueryExpansion schema에 맞게 query expansion하라.
-
-        입력 특약:
-        {clause_text}
-
-        작성 지시:
-        - 출력은 순수 JSON 객체 하나만 생성한다.
-        - 코드블록, 설명문, 주석을 출력하지 않는다.
-        - 최종 법률 판단을 하지 말고 검색용 중간 산출물만 생성한다.
-        - 법령, 판례, 상담사례 검색 질의를 각각 분리한다.
-        - article_no는 문자열 또는 null로 작성한다.
-        - 불확실한 조문 후보는 빈 배열로 둬도 된다.
-        - 자유형 쟁점과 정규화 쟁점을 모두 작성한다.
-        - issue_types_normalized는 허용된 한글 문자열 값만 사용한다.
-        - source별 target_fields는 허용값만 사용한다.
-        - source_routing은 특별한 사유가 없으면 []로 둔다.
-        - source_routing에 confidence 필드를 넣지 않는다.
-        """
-    ).strip()
